chore!: renames non-standard folder, ‘utils’ became 'util'
