+++ src/data/protocols/util/object/index.ts
