+++ src/data/protocols/util/object/process-value.ts
