+++ src/data/protocols/util/text/index.ts
