+++ src/presentation/middlewares/token/validate-token-middleware.ts
@@ -1,4 +1,4 @@
-import { Logger } from '@/data/protocols/utils';
+import { Logger } from '@/data/protocols/util';
 import { ErrorHandler } from '@/domain/usecases';
 import { ValidateToken } from '@/domain/usecases/token';
 import { Middleware } from '@/presentation/protocols/middleware';
