+++ test/util/logger-stub.ts
@@ -1,4 +1,4 @@
-import { Logger } from '@/data/protocols/utils';
+import { Logger } from '@/data/protocols/util';
 
 export class LoggerStub implements Logger {
   log(params: {
